Extract helpers and move tests
